split meal_list.dart into empty_meal_list_element.dart and meal_list_element.dart
